Adding a safe reference counter for initialization/deinitialization (#47)
This counter is necessary when the application makes use of `multiples libraries` each dependent on `libinstpatch`. Examples: 
- `Swami` is dependent on `libinstpatch`and `fluidsynth`. `fluidsynth` could be optionally dependent on `libinstpatch`. In this case, `libinstpatch` is owned twice (by `swami` and by `fluidsynth`).
- For an application only dependent on `fluidsynth`, `libinstpatch` is only owned by `fluidsynth`.
- For an application only dependent on `libinstpatch`, `libinstpatch` is only owned by this application.

This counter ensures that internal `initialization` (or `deinitialization`) is done only one time.
This is required because `libinstpatch` is dependent on `GObject` which doesn't accept that the same `Gtype` be registered multiple times.